Use a replace dictionary that will be compiled from application settings on startup.
The main purpose are message strings which - in most games - have placements of line breaks for the display. Since every message is individual and the line break can therefore occur at different places for each new occurrence of a message, we now let WfcPatcher do the job of generating all possible variations of the message.